move all Webpack config to folder
