+++ .eslintrc
@@ -43,7 +43,7 @@
   "settings": {
     "import/resolver": {
       "webpack": {
-        "config": "webpack.config.eslint.js"
+        "config": "webpack/webpack.config.eslint.js"
       }
     }
   }
