--- flow-typed/npm/module_vx.x.x.js
@@ -1,3 +1,0 @@
-declare module 'module' {
-  declare module.exports: any;
-}
