+++ webpack/webpack.config.base.js
@@ -4,7 +4,7 @@
 
 import path from 'path';
 import webpack from 'webpack';
-import { dependencies as externals } from './app/package.json';
+import { dependencies as externals } from '../app/package.json';
 
 export default {
   externals: Object.keys(externals || {}),
@@ -23,7 +23,7 @@ export default {
   },
 
   output: {
-    path: path.join(__dirname, 'app'),
+    path: path.join(process.cwd(), 'app'),
     filename: 'bundle.js',
     // https://github.com/webpack/webpack/issues/1114
     libraryTarget: 'commonjs2'
@@ -35,7 +35,7 @@ export default {
   resolve: {
     extensions: ['.js', '.jsx', '.json'],
     modules: [
-      path.join(__dirname, 'app'),
+      path.join(process.cwd(), 'app'),
       'node_modules',
     ],
   },
