+++ webpack/webpack.config.renderer.dev.js
@@ -1,28 +1,52 @@
+/* eslint global-require: 0, import/no-dynamic-require: 0 */
+
 /**
- * Builds the DLL for development electron renderer process
+ * Build config for development electron renderer process that uses
+ * Hot-Module-Replacement
+ *
+ * https://webpack.js.org/concepts/hot-module-replacement/
  */
 
-import webpack from 'webpack';
 import path from 'path';
+import fs from 'fs';
+import webpack from 'webpack';
+import chalk from 'chalk';
 import merge from 'webpack-merge';
+import { spawn, execSync } from 'child_process';
+import ExtractTextPlugin from 'extract-text-webpack-plugin';
 import baseConfig from './webpack.config.base';
-import { dependencies } from './package.json';
 
-const dist = path.resolve(process.cwd(), 'dll');
+const port = process.env.PORT || 1212;
+const publicPath = `http://localhost:${port}/dist`;
+const dll = path.resolve(process.cwd(), 'dll');
+const manifest = path.resolve(dll, 'vendor.json');
+
+/**
+ * Warn if the DLL is not built
+ */
+if (!(fs.existsSync(dll) && fs.existsSync(manifest))) {
+  console.log(chalk.black.bgYellow.bold(
+    'The DLL files are missing. Sit back while we build them for you with "npm run build-dll"'
+  ));
+  execSync('npm run build-dll');
+}
 
 export default merge.smart(baseConfig, {
-  context: process.cwd(),
-
-  devtool: 'eval',
+  devtool: 'inline-source-map',
 
   target: 'electron-renderer',
 
-  externals: ['fsevents', 'crypto-browserify'],
+  entry: [
+    'react-hot-loader/patch',
+    `webpack-dev-server/client?http://localhost:${port}/`,
+    'webpack/hot/only-dev-server',
+    path.join(process.cwd(), 'app/index.js'),
+  ],
 
-  /**
-   * @HACK: Copy and pasted from renderer dev config. Consider merging these
-   *        rules into the base config. May cause breaking changes.
-   */
+  output: {
+    publicPath: `http://localhost:${port}/dist/`
+  },
+
   module: {
     rules: [
       {
@@ -152,33 +176,22 @@ export default merge.smart(baseConfig, {
     ]
   },
 
-  resolve: {
-    modules: [
-      'app',
-      'node_modules',
-    ],
-  },
-
-  entry: {
-    vendor: [
-      'babel-polyfill',
-      ...Object.keys(dependencies)
-    ]
-    .filter(dependency => dependency !== 'font-awesome'),
-  },
-
-  output: {
-    library: 'vendor',
-    path: dist,
-    filename: '[name].dll.js',
-    libraryTarget: 'var'
-  },
-
   plugins: [
-    new webpack.DllPlugin({
-      path: path.join(dist, '[name].json'),
-      name: '[name]',
+    new webpack.DllReferencePlugin({
+      context: process.cwd(),
+      manifest: require(manifest),
+      sourceType: 'var',
     }),
+
+    /**
+     * https://webpack.js.org/concepts/hot-module-replacement/
+     */
+    new webpack.HotModuleReplacementPlugin({
+      // @TODO: Waiting on https://github.com/jantimon/html-webpack-plugin/issues/533
+      // multiStep: true
+    }),
+
+    new webpack.NoEmitOnErrorsPlugin(),
 
     /**
      * Create global constants which can be configured at compile time.
@@ -188,19 +201,52 @@ export default merge.smart(baseConfig, {
      *
      * NODE_ENV should be production so that modules do not perform certain
      * development checks
+     *
+     * By default, use 'development' as NODE_ENV. This can be overriden with
+     * 'staging', for example, by changing the ENV variables in the npm scripts
      */
     new webpack.DefinePlugin({
       'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV || 'development')
     }),
 
     new webpack.LoaderOptionsPlugin({
-      debug: true,
-      options: {
-        context: path.resolve(process.cwd(), 'app'),
-        output: {
-          path: path.resolve(process.cwd(), 'dll'),
-        },
-      },
-    })
+      debug: true
+    }),
+
+    new ExtractTextPlugin({
+      filename: '[name].css'
+    }),
   ],
+
+  devServer: {
+    port,
+    publicPath,
+    compress: true,
+    noInfo: true,
+    stats: 'errors-only',
+    inline: true,
+    lazy: false,
+    hot: true,
+    headers: { 'Access-Control-Allow-Origin': '*' },
+    contentBase: path.join(process.cwd(), 'dist'),
+    watchOptions: {
+      aggregateTimeout: 300,
+      poll: 100
+    },
+    historyApiFallback: {
+      verbose: true,
+      disableDotRule: false,
+    },
+    setup() {
+      if (process.env.START_HOT) {
+        spawn(
+          'npm',
+          ['run', 'start-hot-renderer'],
+          { shell: true, env: process.env, stdio: 'inherit' }
+        )
+        .on('close', code => process.exit(code))
+        .on('error', spawnError => console.error(spawnError));
+      }
+    }
+  },
 });
